Update virtual server config to use node list
Prior to this change the virtual server management was using a separate var (bigip_service) to populate the pool . This var overlapped with the node list (bigip_node) that was already used to create the node objects.  So, we can simplify the config by just re-using the node list to fill the pool.